Add another download method to the mix
