+++ src/udemy_dl/udemy_dl.py
@@ -8,13 +8,8 @@ import sys
 import re
 import os
 import json
-import subprocess
 from bs4 import BeautifulSoup
-
-try:
-    from urllib import urlretrieve  # Python 2
-except ImportError:
-    from urllib.request import urlretrieve  # Python 3
+from download import download
 
 
 class Session:
@@ -132,31 +127,14 @@ def mkdir(directory):
         os.makedirs(directory)
 
 
-def dl_progress(num_blocks, block_size, total_size):
-    progress = num_blocks * block_size * 100 / total_size
-    if num_blocks != 0:
-        sys.stdout.write(4 * '\b')
-    sys.stdout.write('%3d%%' % (progress))
-
-
 def get_video(directory, filename, link):
     print('Downloading %s  ' % (filename.encode('utf-8')))
     previous_dir = os.getcwd()
     mkdir(directory)
     os.chdir(directory)
-    try:
-        curl_dl(link, filename)
-    except OSError:
-        if not os.path.exists(filename):
-            urlretrieve(link, filename, reporthook=dl_progress)
-        else:
-            print('Skipping this lecture because an existing file already exists')
+    download(link, filename)
     os.chdir(previous_dir)
     print('\n'),
-
-def curl_dl(link, filename):
-    command = ['curl', '-C', '-', link, '-o', filename ,'--insecure']
-    subprocess.call(command)
 
 def udemy_dl(username, password, course_link, dest_dir=""):
     login(username, password)
